move colour logic to component
